--- .env.production
@@ -1,1 +1,0 @@
-VITE_APP_ORIGIN=https://pqsoft.net
