+++ package.json
@@ -19,7 +19,8 @@
     "react-hook-form": "^7.51.3",
     "react-router-dom": "^6.22.3",
     "socket.io-client": "^4.7.5",
-    "styled-components": "^6.1.8"
+    "styled-components": "^6.1.8",
+    "zustand": "^4.5.2"
   },
   "devDependencies": {
     "@tanstack/eslint-plugin-query": "^5.28.11",
