+++ src/apis/instance/axiosInstance.ts
@@ -1,7 +1,12 @@
 import axios from 'axios';
-import Cookies from 'js-cookie';
 import { URL } from 'src/constants/apiUrl';
+import useUserStore from 'src/store/userStore';
 import handleError from 'src/utils/handleError';
+
+const getAccessToken = () => {
+  const store = useUserStore.getState();
+  return store.accessToken;
+};
 
 /**
  * axios instance 속성 추가
@@ -16,7 +21,7 @@ const axiosInstance = axios.create({
 
 axiosInstance.interceptors.request.use(
   (config) => {
-    const accessToken = Cookies.get('accessToken');
+    const accessToken = getAccessToken();
     if (accessToken) {
       config.headers['Authorization'] = `Bearer ${accessToken}`;
     }
