+++ src/pages/signup/_components/EmailCheckForm.tsx
@@ -4,7 +4,11 @@ import { SubmitButton } from 'src/components/sign/button/SignSubmitButton';
 import { ButtonNormal } from 'src/GlobalStyles';
 import AuthInput from 'src/components/sign/input/AuthInput';
 import { FormValues } from '../_types/type';
-import { useCheckEmail } from 'src/hooks/useCheckEmail';
+import { useMutationPost } from 'src/apis/service/service';
+import { URL } from 'src/constants/apiUrl';
+import { AxiosError } from 'axios';
+import { ERROR_MESSAGES } from 'src/constants/error';
+import Cookies from 'js-cookie';
 
 export default function EmailCheckForm() {
   const {
@@ -16,18 +20,42 @@ export default function EmailCheckForm() {
     formState: { errors },
   } = useForm<FormValues>();
 
-  const { mutate, isPending } = useCheckEmail(setError);
+  const { mutate, isPending } = useMutationPost(`${URL.AUTH}/signup/confirm`, {
+    onError: (error: unknown) => {
+      const axiosError = error as AxiosError;
+      const status = axiosError?.response?.status;
 
-  const onSubmit = async (data: FormValues) => {
+      if (status === 400) {
+        setError('otp', {
+          type: 'custom',
+          message: ERROR_MESSAGES.AUTH.EMAIL_VERIFY_CHECK_FAILED,
+        });
+
+        return;
+      }
+
+      setError('otp', {
+        type: 'custom',
+        message: ERROR_MESSAGES.AUTH.EMAIL_VERIFY_FAILED,
+      });
+      alert(ERROR_MESSAGES.AUTH.EMAIL_VERIFY_FAILED);
+    },
+
+    onSuccess: () => {
+      location.replace('/login');
+    },
+  });
+
+  const onSubmit = (data: FormValues) => {
     if (isPending) {
       return;
     }
 
     const verificationCode = Object.values(data).join('');
-    const email = localStorage.getItem('email'); // 임시
+    const email = Cookies.get('email');
 
     const EmailVerifyData = {
-      email: email as string,
+      email: email,
       code: verificationCode,
     };
